requirements.txt
esdk-obs-python==3.23.9.1
pycryptodome==3.10.1
